fixed error with delete all data and then add product + validation on input in form + if yiu edit product from list level you cannot delete product

+++ Projekt/frontend/src/Component/form.js
@@ -36,9 +36,19 @@ class Form extends Component {
                     id: 0,
                     accept: false,
                     email: "",
-                    content: ""
+                    content: "",
+                    validation: {
+                        email: false,
+                        content: false
+                    }
                 }
-            ]
+            ],
+            validation: {
+                name: false,
+                description: false,
+            },
+            deleteAll: false,
+            delete: false
         }
 
         this.takeProduct = this.takeProduct.bind(this)
@@ -66,8 +76,18 @@ class Form extends Component {
                     id: take_product.comments.length,
                     accept: false,
                     email: "",
-                    content: ""
-                }]
+                    content: "",
+                    validation: {
+                        email: false,
+                        content: false
+                    }
+                }],
+            validation: {
+                name: false,
+                description: false,
+            },
+            deleteAll: false,
+            delete: false
         })     
     }
 
@@ -79,8 +99,8 @@ class Form extends Component {
     }
 
     createInput(word, what){
-        if(what !== 'product_colors' ){
-            return ((what !== 'description' && what !== 'id' && what !== 'active' &&
+        if(what !== 'product_colors' && what !== 'comments' ){
+            return ((what !== 'description' && what !== 'id' && what !== 'active' && what !== 'name' &&
                 <div key={what}>
                     <b>{word}</b><br/>
                     <input
@@ -88,17 +108,26 @@ class Form extends Component {
                         onChange={(ev) => this.handleChange(ev.target.value, what)}
                         value={this.state.product[what]} /><br/>
                 </div>
-                ) || (what === 'description' && what !== 'id' && what !== 'active' &&
+                ) || (what === 'description' &&
                 <div key={what}>
                     <b>{word}</b><br/>
                     <textarea 
                         className="description"
                         onChange={(ev) => this.handleChange(ev.target.value, 'description')} 
                         value={this.state.product.description} /><br/>
+                    {this.state.validation.description && (<p className="error">At least 100 letters in description</p>)}
                 </div>
-            ))
-        }
+                ) || (what === 'name' && 
+                <div key={what}>
+                    <b>{word}</b><br/>
+                    <input
+                        type="text"
+                        onChange={(ev) => this.handleChange(ev.target.value, 'name')} 
+                        value={this.state.product.name} /><br/>
+                    {this.state.validation.name && (<p className="error">Name is required</p>)}
+                </div>))}   
     }
+    
 
     handleDone(id){
         const product = this.state.product
@@ -114,15 +143,35 @@ class Form extends Component {
         this.setState({...this.state, product: {...product}})
 
 
-
         if(this.state.product.name.length === 0){
-            alert("Name is required")
+            if(this.state.product.description.split("").length < 100){
+                this.setState({...this.state, 
+                    validation: {
+                        name: true,
+                        description: true
+                    }
+                })
+            }
+            else{
+                this.setState({...this.state, 
+                    validation: {
+                        name: true,
+                        description: false
+                    }
+                })
+            }
         }
         else if(this.state.product.description.split("").length < 100){
-            alert("At least 100 letters in description")
+            this.setState({...this.state, 
+                validation: {
+                    name: false,
+                    description: true
+                }
+            })
         }
         else{
-            if(this.state.take !== 'new-product'){                
+            if(this.state.take !== 'new-product'){
+                this.props.editDone(this.state.product.id, {...this.state.product})             
                 this.setState({
                     take: "new-product",
                     product: prepare_product,
@@ -131,15 +180,26 @@ class Form extends Component {
                             id: 0,
                             accept: false,
                             email: "",
-                            content: ""
+                            content: "",
+                            validation: {
+                                email: false,
+                                content: false
+                            }
                         }
-                    ]
+                    ],
+                    validation: {
+                        name: false,
+                        description: false,
+                    },
+                    deleteAll: false,
+                    delete: false
                 })
             }
             else{
                 const filterComment = this.state.comments.filter(item => item.accept !== false)
                 const new_product = this.state.product
                 new_product['id'] = id + 1
+                console.log(new_product.id)
                 this.props.newProduct({...new_product, comments: filterComment})
                 this.setState({
                     ...this.state,
@@ -149,7 +209,11 @@ class Form extends Component {
                             id: 0,
                             accept: false,
                             email: "",
-                            content: ""
+                            content: "",
+                            validation: {
+                                email: false,
+                                content: false
+                            }
                         }
                     ]
                 })
@@ -162,13 +226,71 @@ class Form extends Component {
         if(this.state.take !== 'new-product'){
             this.props.deleteProduct(this.state.product.id)
             this.setState({
-                ...this.state, product: prepare_product
+                take: "new-product",
+                product: prepare_product,
+                comments: [
+                    {
+                        id: 0,
+                        accept: false,
+                        email: "",
+                        content: "",
+                        validation: {
+                            email: false,
+                            content: false
+                        }
+                    }
+                ],
+                validation: {
+                    name: false,
+                    description: false,
+                },
+                deleteAll: false,
+                delete: true
             })
+
+            setTimeout(() => {
+                this.setState({...this.state,
+                    delete: false
+                })
+            }, 1000)
         }
     }   
 
     handleDeleteAll(){
-        this.props.deleteAllProducts()
+        let result = window.confirm("Do you want delete all data?")
+        if (result === true){
+            this.props.deleteAllProducts()
+            this.setState({take: "new-product",
+            product: prepare_product,
+            comments: [
+                {
+                    id: 0,
+                    accept: false,
+                    email: "",
+                    content: "",
+                    validation: {
+                        email: false,
+                        content: false
+                    }
+                }
+            ],
+            validation: {
+                name: false,
+                description: false,
+            },
+            deleteAll: true,
+            delete: false
+            })
+
+            setTimeout(() => {
+                this.setState({...this.state,
+                    deleteAll: false
+                })
+            }, 3000)
+
+        }
+
+
     }
 
     handleReset(){
@@ -180,25 +302,57 @@ class Form extends Component {
                     id: 0,
                     accept: false,
                     email: "",
-                    content: ""
+                    content: "",
+                    validation: {
+                        email: false,
+                        content: false
+                    }
                 }
-            ]
+            ],
+            validation: {
+                name: false,
+                description: false,
+            },
+            deleteAll: false,
+            delete: false
         })
     }
 
     handelAddComment(){
         const re = /\S+@\S+\.\S+/
         const id = this.state.comments[this.state.comments.length - 1].id + 1
+        const com = this.state.comments
         if(!re.test(this.state.comments[id - 1].email)){
-            alert("Wrong email")
+            if(this.state.comments[id - 1].content.length === 0){
+                com[this.state.comments.length - 1].validation.email = true
+                com[this.state.comments.length - 1].validation.content = true
+                this.setState({...this.state, 
+                    comments: com
+                })
+            }
+            else{
+                com[this.state.comments.length - 1].validation.email = true
+                com[this.state.comments.length - 1].validation.content = false
+
+                this.setState({...this.state, 
+                    comments: com
+                })
+            }
         }
         else if(this.state.comments[id - 1].content.length === 0){
-            alert("Empty content")
+            com[this.state.comments.length - 1].validation.email = false
+            com[this.state.comments.length - 1].validation.content = true
+
+            this.setState({...this.state, 
+                comments: com
+            })
         }
-        else{
+        else{ 
             const upload = this.state.comments.map(item => {
                 if(item.id === id - 1){
                     item.accept = true;
+                    item.validation.email = false;
+                    item.validation.content = false;
                     return item
                 }
                 return item
@@ -215,7 +369,11 @@ class Form extends Component {
                         id: id,
                         accept: false,
                         email: "",
-                        content: ""
+                        content: "",
+                        validation: {
+                            email: false,
+                            content: false
+                        }
                     }
                 ]
             })  
@@ -243,6 +401,12 @@ class Form extends Component {
 
         return(
             <div className="form">
+                {this.state.deleteAll && (
+                    <div className="confirmDeleteAll">You delete all data. Now you can start create your own products!</div>
+                )}
+                {this.state.delete && (
+                    <div className="confirmDeleteAll">Deleted!!!</div>
+                )}
                 <div className="formEdit">
                     <select
                         onChange={(ev) => this.takeProduct(ev.target.value)}
@@ -278,14 +442,14 @@ class Form extends Component {
                         <input 
                             type="button"
                             value="Done"
-                            onClick={() => this.handleDone(data[0].id)}/>
+                            onClick={() => this.handleDone(data.length > 0 ? data[0].id : 0)}/>
                         <input
                             type="button"
-                            value="Delete"
+                            value="Delete this product"
                             onClick={() => this.handleDelete()}/>
                         <input
                             type="button"
-                            value="Delete all"
+                            value="Delete all products"
                             onClick={() => this.handleDeleteAll()}/>
                         <input 
                             type="button"
@@ -311,10 +475,12 @@ class Form extends Component {
                                 value={comment.email}
                                 onChange={(ev) => !comment.accept && this.handleChangeComment(ev.target.value, 'email', comment.id)}
                                 placeholder="Email"/>
+                            {comment.validation.email && (<p className="error">Wrong email</p>)}
                             <textarea
                                 placeholder="Content"
                                 onChange={(ev) => !comment.accept && this.handleChangeComment(ev.target.value, 'content', comment.id)} 
                                 value={comment.content}/>
+                            {comment.validation.content && (<p className="error">Empty content</p>)}
                             {!comment.accept && (
                             <input 
                                 className="addComment"
